// src/api/generated/models/BranchResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BranchResponse = {
    id?: string;
    workspaceId?: string;
    name: string;
    description?: string;
    headCommitId?: string;
    baseCommitId?: string;
    isDefault: boolean;
    isArchived: boolean;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
};